# document/spec/openapi.yaml
openapi: 3.1.0
info:
  title: lib4gz LMS Platform Backend API
  description: >
    RESTful API for the lib4gz Learning Management System platform.
    Provides endpoints for user authentication, course management, enrollments,
    modules, lessons, summaries, exercises, questions, and submissions.
  version: 1.0.0
  contact:
    name: lib4gz Team
servers:
  - url: http://localhost:8080/api
    description: Local Development
security:
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: access_token
      description: JWT access token
  parameters:
    XUserIdHeader:
      name: X-User-Id
      in: header
      required: true
      description: TypeID of the authenticated user, injected by JwtAuthFilter
      schema:
        type: string
        format: typeid
    CourseIdPath:
      name: courseId
      in: path
      required: true
      description: TypeID of the course
      schema:
        type: string
        format: typeid
    ModuleIdPath:
      name: moduleId
      in: path
      required: true
      description: TypeID of the module
      schema:
        type: string
        format: typeid
    LessonIdPath:
      name: lessonId
      in: path
      required: true
      description: TypeID of the lesson
      schema:
        type: string
        format: typeid
    ExerciseIdPath:
      name: exerciseId
      in: path
      required: true
      description: TypeID of the exercise
      schema:
        type: string
        format: typeid
...
  schemas:
...
    ExerciseType:
      type: string
      enum:
        - TEXT_ANSWER
        - CODE
        - FILE_UPLOAD
        - MULTIPLE_CHOICE
        - PROJECT_LINK
...
    UserSummary:
      type: object
      properties:
        id:
          type: string
          format: typeid
        name:
          type: string
        email:
          type: string
        avatarUrl:
          type:
            - string
            - "null"

    # ---------- Courses ----------
...
    CreateModuleRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        orderIndex:
          type:
            - integer
            - "null"
    UpdateModuleRequest:
      type: object
      properties:
        title:
          type:
            - string
            - "null"
        orderIndex:
          type:
            - integer
            - "null"
    ModuleResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        courseId:
          type: string
          format: typeid
        title:
          type: string
        orderIndex:
          type: integer
        lessonCount:
          type:
            - integer
            - "null"
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64

    # ---------- Lessons ----------
    CreateLessonRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        orderIndex:
          type:
            - integer
            - "null"
    UpdateLessonRequest:
      type: object
      properties:
        title:
          type:
            - string
            - "null"
        orderIndex:
          type:
            - integer
            - "null"
    LessonResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        moduleId:
          type: string
          format: typeid
        title:
          type: string
        orderIndex:
          type: integer
        hasSummary:
          type: boolean
        exerciseCount:
          type:
            - integer
            - "null"
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64

    # ---------- Summaries ----------
    CreateSummaryRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
...
    SummaryResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        lessonId:
          type: string
          format: typeid
        content:
          type: string
        editedBy:
          $ref: "#/components/schemas/UserSummary"
        version:
          type: integer
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64

    # ---------- Exercises ----------
    CreateExerciseRequest:
      type: object
      required:
        - title
        - type
      properties:
        title:
          type: string
        type:
          $ref: "#/components/schemas/ExerciseType"
        settings:
          type: object
          default: {}
        orderIndex:
          type:
            - integer
            - "null"
    UpdateExerciseRequest:
      type: object
      properties:
        title:
          type:
            - string
            - "null"
        type:
          oneOf:
            - $ref: "#/components/schemas/ExerciseType"
            - type: "null"
        settings:
          type:
            - object
            - "null"
        orderIndex:
          type:
            - integer
            - "null"
    ExerciseResponse:
      type: object
      properties:
        id:
          type: string
          format: typeid
        lessonId:
          type: string
          format: typeid
        title:
          type: string
        type:
          $ref: "#/components/schemas/ExerciseType"
        settings:
          type: object
        orderIndex:
          type: integer
        questionCount:
          type:
            - integer
            - "null"
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64

    # ---------- Questions ----------
...
    ErrorResp:
      type: object
      properties:
        timestamp:
          type: integer
          format: int64
        path:
          type: string
        status:
          type: integer
        requestId:
          type: string
        traceId:
          type: string
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ApiError"
    ApiError:
      oneOf:
        - $ref: "#/components/schemas/InputError"
        - $ref: "#/components/schemas/GenericError"
      discriminator:
        propertyName: type
    InputError:
      type: object
      properties:
        type:
          type: string
          enum:
            - InputError
        field:
          type: string
        message:
          type: string
    GenericError:
      type: object
      properties:
        type:
          type: string
          enum:
            - SystemError
        reason:
          type: string
  responses:
    BadRequest:
      description: Bad request - validation errors or malformed input
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    Unauthorized:
      description: Unauthorized - missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResp"
paths:
...
  /v1/courses/{courseId}/modules:
    post:
      tags:
        - modules
      operationId: createModule
      summary: Create a new module in a course
      description: Creates a new module within the specified course. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/CourseIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleRequest"
      responses:
        "201":
          description: Module created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    get:
      tags:
        - modules
      operationId: listModules
      summary: List modules in a course
      description: Retrieves all modules for the specified course, ordered by orderIndex.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/CourseIdPath"
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ModuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/modules/{moduleId}:
    get:
      tags:
        - modules
      operationId: getModule
      summary: Get a module by ID
      description: Retrieves the details of a specific module.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ModuleIdPath"
      responses:
        "200":
          description: Module details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - modules
      operationId: updateModule
      summary: Update a module
      description: Partially updates a module. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ModuleIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModuleRequest"
      responses:
        "200":
          description: Module updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModuleResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - modules
      operationId: deleteModule
      summary: Delete a module
      description: Deletes a module and all its lessons. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ModuleIdPath"
      responses:
        "204":
          description: Module deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # LESSONS
  # ========================================================================
  /v1/modules/{moduleId}/lessons:
    post:
      tags:
        - lessons
      operationId: createLesson
      summary: Create a new lesson in a module
      description: Creates a new lesson within the specified module. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ModuleIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonRequest"
      responses:
        "201":
          description: Lesson created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    get:
      tags:
        - lessons
      operationId: listLessons
      summary: List lessons in a module
      description: Retrieves all lessons for the specified module, ordered by orderIndex.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ModuleIdPath"
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LessonResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/lessons/{lessonId}:
    get:
      tags:
        - lessons
      operationId: getLesson
      summary: Get a lesson by ID
      description: Retrieves the details of a specific lesson.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      responses:
        "200":
          description: Lesson details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - lessons
      operationId: updateLesson
      summary: Update a lesson
      description: Partially updates a lesson. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLessonRequest"
      responses:
        "200":
          description: Lesson updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - lessons
      operationId: deleteLesson
      summary: Delete a lesson
      description: Deletes a lesson and all its content. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      responses:
        "204":
          description: Lesson deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # SUMMARIES
  # ========================================================================
  /v1/lessons/{lessonId}/summary:
    get:
      tags:
        - summaries
      operationId: getLessonSummary
      summary: Get the summary for a lesson
      description: Retrieves the summary content for the specified lesson.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      responses:
        "200":
          description: Summary details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    post:
      tags:
        - summaries
      operationId: createOrUpdateLessonSummary
      summary: Create or update a lesson summary
      description: >
        Creates a new summary for the lesson if none exists, or updates the existing summary.
        Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSummaryRequest"
      responses:
        "201":
          description: Summary created or updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SummaryResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # EXERCISES
  # ========================================================================
  /v1/lessons/{lessonId}/exercises:
    post:
      tags:
        - exercises
      operationId: createExercise
      summary: Create a new exercise in a lesson
      description: Creates a new exercise within the specified lesson. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExerciseRequest"
      responses:
        "201":
          description: Exercise created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    get:
      tags:
        - exercises
      operationId: listExercises
      summary: List exercises in a lesson
      description: Retrieves all exercises for the specified lesson, ordered by orderIndex.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/LessonIdPath"
      responses:
        "200":
          description: List of exercises
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExerciseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/exercises/{exerciseId}:
    get:
      tags:
        - exercises
      operationId: getExercise
      summary: Get an exercise by ID
      description: Retrieves the details of a specific exercise.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      responses:
        "200":
          description: Exercise details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      tags:
        - exercises
      operationId: updateExercise
      summary: Update an exercise
      description: Partially updates an exercise. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateExerciseRequest"
      responses:
        "200":
          description: Exercise updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExerciseResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags:
        - exercises
      operationId: deleteExercise
      summary: Delete an exercise
      description: Deletes an exercise and all its questions and submissions. Restricted to teachers and course creators.
      parameters:
        - $ref: "#/components/parameters/XUserIdHeader"
        - $ref: "#/components/parameters/ExerciseIdPath"
      responses:
        "204":
          description: Exercise deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ========================================================================
  # QUESTIONS
  # ========================================================================
...